Adjustments
- Adjust repository layout
- include Glitch Archive
