+++ SOCOM II/Acid Trip/Acid Trip.txt
@@ -1,0 +1,144 @@
+Controls
+-----------------
+L1+Left  :  Add Red
+L1+Right  :  Subtract Red
+L2+Left  :  Add Green
+L2+Right  :  Subtract Green
+R2+Left  :  Add Blue
+R2+Right  :  Subtract Blue
+L3+R3  :  Reset Colors
+Notes: this is for R1 patch. Does not work online. Can be ported to use online if anyone’s interested.
+-----------------
+Acid Trip
+202CED04 08039278
+200E49E0 27BDFFE0
+200E49E4 3C080045
+200E49E8 AFBF0000
+200E49EC 3C09000E
+200E49F0 AFB00004
+200E49F4 AFB10008
+200E49F8 AFB2000C
+200E49FC 0C0ACD60
+200E4A00 00000000
+200E4A04 0040802D
+200E4A08 1200004E
+200E4A0C 00000000
+200E4A10 813249D0
+200E4A14 24030010
+200E4A18 12430005
+200E4A1C 00000000
+200E4A20 26520001
+200E4A24 A13249D0
+200E4A28 10000046
+200E4A2C 00000000
+200E4A30 A12049D0
+200E4A34 8511F15C
+200E4A38 2402FB7F
+200E4A3C 12220015
+200E4A40 00000000
+200E4A44 2402FBDF
+200E4A48 12220018
+200E4A4C 00000000
+200E4A50 2402FE7F
+200E4A54 1222001B
+200E4A58 00000000
+200E4A5C 2402FEDF
+200E4A60 1222001E
+200E4A64 00000000
+200E4A68 2402FD7F
+200E4A6C 12220021
+200E4A70 00000000
+200E4A74 2402FDDF
+200E4A78 12220024
+200E4A7C 00000000
+200E4A80 2402FFF9
+200E4A84 12220027
+200E4A88 00000000
+200E4A8C 1000002D
+200E4A90 00000000
+200E4A94 3C04004B
+200E4A98 34844D40
+200E4A9C 0C0392D8
+200E4AA0 00000000
+200E4AA4 10000027
+200E4AA8 00000000
+200E4AAC 3C04004B
+200E4AB0 34844D40
+200E4AB4 0C0392E2
+200E4AB8 00000000
+200E4ABC 10000021
+200E4AC0 00000000
+200E4AC4 3C04004B
+200E4AC8 34844D44
+200E4ACC 0C0392D8
+200E4AD0 00000000
+200E4AD4 1000001B
+200E4AD8 00000000
+200E4ADC 3C04004B
+200E4AE0 34844D44
+200E4AE4 0C0392E2
+200E4AE8 00000000
+200E4AEC 10000015
+200E4AF0 00000000
+200E4AF4 3C04004B
+200E4AF8 34844D48
+200E4AFC 0C0392D8
+200E4B00 00000000
+200E4B04 1000000F
+200E4B08 00000000
+200E4B0C 3C04004B
+200E4B10 34844D48
+200E4B14 0C0392E2
+200E4B18 00000000
+200E4B1C 10000009
+200E4B20 00000000
+200E4B24 3C04004B
+200E4B28 34844D40
+200E4B2C 3C05004B
+200E4B30 34A54D44
+200E4B34 3C06004B
+200E4B38 34C64D48
+200E4B3C 0C0392EC
+200E4B40 00000000
+200E4B44 8FBF0000
+200E4B48 8FB00004
+200E4B4C 8FB10008
+200E4B50 8FB2000C
+200E4B54 03E00008
+200E4B58 27BD0020
+200E4B5C 00000000
+200E4B60 3C023F80
+200E4B64 44820800
+200E4B68 C4800000
+200E4B6C 46000880
+200E4B70 E4820000
+200E4B74 E4820010
+200E4B78 E4820020
+200E4B7C E4820030
+200E4B80 03E00008
+200E4B84 00000000
+200E4B88 3C023F80
+200E4B8C 44820800
+200E4B90 C4800000
+200E4B94 46010081
+200E4B98 E4820000
+200E4B9C E4820010
+200E4BA0 E4820020
+200E4BA4 E4820030
+200E4BA8 03E00008
+200E4BAC 00000000
+200E4BB0 3C023F80
+200E4BB4 44820800
+200E4BB8 E4810000
+200E4BBC E4810010
+200E4BC0 E4810020
+200E4BC4 E4810030
+200E4BC8 E4A10000
+200E4BCC E4A10010
+200E4BD0 E4A10020
+200E4BD4 E4A10030
+200E4BD8 E4C10000
+200E4BDC E4C10010
+200E4BE0 E4C10020
+200E4BE4 E4C10030
+200E4BE8 03E00008
