+++ SOCOM II/Aimbot/r0001/inSOCOMenu.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31205.134
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "inSOCOMenu", "inSOCOMenu.vcxproj", "{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Debug|x64.ActiveCfg = Debug|x64
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Debug|x64.Build.0 = Debug|x64
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Debug|x86.ActiveCfg = Debug|Win32
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Debug|x86.Build.0 = Debug|Win32
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Release|x64.ActiveCfg = Release|x64
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Release|x64.Build.0 = Release|x64
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Release|x86.ActiveCfg = Release|Win32
+		{4799C5CA-C7E7-4F0B-AFB5-BE429C2470A8}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {16FD4FAD-1130-485B-91D5-74AD14CD7334}
+	EndGlobalSection
+EndGlobal
